Create seperate knowledge fastapi dependencies module
